chore: reorganize the screenshot method
