+++ src/lib/maps/leaflet/leafletmap.js
@@ -3,9 +3,9 @@ import { LeafletMap } from './basic'
 /**
  * Sub class for rendering web maps with Leaflet.TileLayer.
  */
-class WebMapLeaflet extends LeafletMap {
+class LeafletWebMap extends LeafletMap {
   /**
-   * WebMapLeaflet constructor parameters
+   * LeafletWebMap constructor parameters
    * @typedef {Object} config
    * @param {String} config.baseMapUrl - Basemap URL.
    * @param {String} config.baseMapAttrib - Basemap attribution text.
@@ -31,4 +31,4 @@ class WebMapLeaflet extends LeafletMap {
   }
 }
 
-export default WebMapLeaflet
+export default LeafletWebMap
